Delete ConditionalOrderGrantedNotification.java, add Optional.empty() to document pack, update events
