+++ src/main/java/uk/gov/hmcts/divorce/common/notification/ConditionalOrderPronouncedNotification.java
@@ -1,26 +1,25 @@
 package uk.gov.hmcts.divorce.common.notification;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.divorce.divorcecase.model.Applicant;
 import uk.gov.hmcts.divorce.divorcecase.model.CaseData;
+import uk.gov.hmcts.divorce.document.print.LetterPrinter;
+import uk.gov.hmcts.divorce.document.print.documentpack.ConditionalOrderGrantedDocumentPack;
 import uk.gov.hmcts.divorce.divorcecase.model.ConditionalOrder;
 import uk.gov.hmcts.divorce.notification.ApplicantNotification;
 import uk.gov.hmcts.divorce.notification.CommonContent;
 import uk.gov.hmcts.divorce.notification.EmailTemplateName;
 import uk.gov.hmcts.divorce.notification.NotificationService;
 import uk.gov.hmcts.divorce.notification.exception.NotificationTemplateException;
-import uk.gov.hmcts.divorce.systemupdate.service.print.ConditionalOrderPronouncedPrinter;
 
 import java.time.format.DateTimeFormatter;
 import java.util.Map;
 
 import static java.lang.String.format;
 import static java.util.Objects.isNull;
-import static uk.gov.hmcts.divorce.document.model.DocumentType.CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_1;
-import static uk.gov.hmcts.divorce.document.model.DocumentType.CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_2;
 import static uk.gov.hmcts.divorce.notification.CommonContent.APPLICANT;
 import static uk.gov.hmcts.divorce.notification.CommonContent.APPLICANT1_LABEL;
 import static uk.gov.hmcts.divorce.notification.CommonContent.APPLICANT2_LABEL;
@@ -39,6 +38,7 @@ import static uk.gov.hmcts.divorce.notification.FormatUtil.DATE_TIME_FORMATTER;
 import static uk.gov.hmcts.divorce.notification.FormatUtil.getDateTimeFormatterForPreferredLanguage;
 
 @Component
+@RequiredArgsConstructor
 @Slf4j
 public class ConditionalOrderPronouncedNotification implements ApplicantNotification {
 
@@ -50,14 +50,12 @@ public class ConditionalOrderPronouncedNotification implements ApplicantNotifica
     @Value("${final_order.respondent_eligible_from_offset_months}")
     private long finalOrderRespondentOffsetMonth;
 
-    @Autowired
     private NotificationService notificationService;
 
-    @Autowired
     private CommonContent commonContent;
 
-    @Autowired
-    private ConditionalOrderPronouncedPrinter conditionalOrderPronouncedPrinter;
+    private final ConditionalOrderGrantedDocumentPack conditionalOrderGrantedDocumentPack;
+    private final LetterPrinter letterPrinter;
 
     @Override
     public void sendToApplicant1(final CaseData caseData, final Long caseId) {
@@ -88,12 +86,9 @@ public class ConditionalOrderPronouncedNotification implements ApplicantNotifica
     @Override
     public void sendToApplicant1Offline(final CaseData caseData, final Long caseId) {
         log.info("Sending conditional order letter to applicant 1 for case: {}", caseId);
-        conditionalOrderPronouncedPrinter.sendLetter(
-            caseData,
-            caseId,
-            CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_1,
-            caseData.getApplicant1()
-        );
+        Applicant applicant1 = caseData.getApplicant1();
+        var documentPackInfo = conditionalOrderGrantedDocumentPack.getDocumentPack(caseData, applicant1);
+        letterPrinter.sendLetters(caseData, caseId, applicant1, documentPackInfo, conditionalOrderGrantedDocumentPack.getLetterId());
     }
 
     @Override
@@ -128,12 +123,9 @@ public class ConditionalOrderPronouncedNotification implements ApplicantNotifica
     @Override
     public void sendToApplicant2Offline(final CaseData caseData, final Long caseId) {
         log.info("Sending conditional order letter to applicant 2 for case: {}", caseId);
-        conditionalOrderPronouncedPrinter.sendLetter(
-            caseData,
-            caseId,
-            CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_2,
-            caseData.getApplicant2()
-        );
+        Applicant applicant2 = caseData.getApplicant2();
+        var documentPackInfo = conditionalOrderGrantedDocumentPack.getDocumentPack(caseData, applicant2);
+        letterPrinter.sendLetters(caseData, caseId, applicant2, documentPackInfo, conditionalOrderGrantedDocumentPack.getLetterId());
     }
 
     private Map<String, String> templateVars(final CaseData caseData,
