+++ src/main/java/uk/gov/hmcts/divorce/document/print/documentpack/ConditionalOrderGrantedDocumentPack.java
@@ -20,7 +20,8 @@ public class ConditionalOrderGrantedDocumentPack implements DocumentPack {
     private static final DocumentPackInfo APPLICANT_1_JUDICIAL_SEPARATION_CONDITIONAL_ORDER_PACK = new DocumentPackInfo(
         ImmutableMap.of(
             CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_1, Optional.of(JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID),
-            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID)
+            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID),
+            CONDITIONAL_ORDER_GRANTED, Optional.empty()
         ),
         ImmutableMap.of(
             JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID, JUDICIAL_SEPARATION_ORDER_GRANTED_COVERSHEET_DOCUMENT_NAME,
@@ -31,7 +32,8 @@ public class ConditionalOrderGrantedDocumentPack implements DocumentPack {
     private static final DocumentPackInfo APPLICANT_2_JUDICIAL_SEPARATION_CONDITIONAL_ORDER_PACK = new DocumentPackInfo(
         ImmutableMap.of(
             CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_2, Optional.of(JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID),
-            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID)
+            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID),
+            CONDITIONAL_ORDER_GRANTED, Optional.empty()
         ),
         ImmutableMap.of(
             JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID, JUDICIAL_SEPARATION_ORDER_GRANTED_COVERSHEET_DOCUMENT_NAME,
@@ -42,7 +44,8 @@ public class ConditionalOrderGrantedDocumentPack implements DocumentPack {
     private static final DocumentPackInfo APPLICANT_1_SOL_JUDICIAL_SEPARATION_CONDITIONAL_ORDER_PACK = new DocumentPackInfo(
         ImmutableMap.of(
             CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_1, Optional.of(JUDICIAL_SEPARATION_ORDER_GRANTED_SOLICITOR_COVER_LETTER_TEMPLATE_ID),
-            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID)
+            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID),
+            CONDITIONAL_ORDER_GRANTED, Optional.empty()
         ),
         ImmutableMap.of(
             JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID, JUDICIAL_SEPARATION_ORDER_GRANTED_SOLICITOR_COVERSHEET_DOCUMENT_NAME,
@@ -53,7 +56,8 @@ public class ConditionalOrderGrantedDocumentPack implements DocumentPack {
     private static final DocumentPackInfo APPLICANT_2_SOL_JUDICIAL_SEPARATION_CONDITIONAL_ORDER_PACK = new DocumentPackInfo(
         ImmutableMap.of(
             CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_2, Optional.of(JUDICIAL_SEPARATION_ORDER_GRANTED_SOLICITOR_COVER_LETTER_TEMPLATE_ID),
-            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID)
+            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID),
+            CONDITIONAL_ORDER_GRANTED, Optional.empty()
         ),
         ImmutableMap.of(
             JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID, JUDICIAL_SEPARATION_ORDER_GRANTED_SOLICITOR_COVERSHEET_DOCUMENT_NAME,
@@ -64,7 +68,8 @@ public class ConditionalOrderGrantedDocumentPack implements DocumentPack {
     private static final DocumentPackInfo APPLICANT_1_SEPARATION_CONDITIONAL_ORDER_PACK = new DocumentPackInfo(
         ImmutableMap.of(
             CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_1, Optional.of(JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID),
-            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID)
+            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID),
+            CONDITIONAL_ORDER_GRANTED, Optional.empty()
         ),
         ImmutableMap.of(
             CO_GRANTED_COVER_LETTER_TEMPLATE_ID, SEPARATION_ORDER_GRANTED_COVERSHEET_DOCUMENT_NAME,
@@ -75,7 +80,8 @@ public class ConditionalOrderGrantedDocumentPack implements DocumentPack {
     private static final DocumentPackInfo APPLICANT_2_SEPARATION_CONDITIONAL_ORDER_PACK = new DocumentPackInfo(
         ImmutableMap.of(
             CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_2, Optional.of(JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID),
-            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID)
+            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID),
+            CONDITIONAL_ORDER_GRANTED, Optional.empty()
         ),
         ImmutableMap.of(
             JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID, SEPARATION_ORDER_GRANTED_COVERSHEET_DOCUMENT_NAME,
@@ -86,7 +92,8 @@ public class ConditionalOrderGrantedDocumentPack implements DocumentPack {
     private static final DocumentPackInfo APPLICANT_1_SOL_SEPARATION_CONDITIONAL_ORDER_PACK = new DocumentPackInfo(
         ImmutableMap.of(
             CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_1, Optional.of(JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID),
-            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID)
+            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID),
+            CONDITIONAL_ORDER_GRANTED, Optional.empty()
         ),
         ImmutableMap.of(
             JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID, SEPARATION_ORDER_GRANTED_SOLICITOR_COVERSHEET_DOCUMENT_NAME,
@@ -97,7 +104,8 @@ public class ConditionalOrderGrantedDocumentPack implements DocumentPack {
     private static final DocumentPackInfo APPLICANT_2_SOL_SEPARATION_CONDITIONAL_ORDER_PACK = new DocumentPackInfo(
         ImmutableMap.of(
             CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_2, Optional.of(JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID),
-            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID)
+            CONDITIONAL_ORDER_GRANTED, Optional.of(JUDICIAL_SEPARATION_ORDER_PRONOUNCED_TEMPLATE_ID),
+            CONDITIONAL_ORDER_GRANTED, Optional.empty()
         ),
         ImmutableMap.of(
             JUDICIAL_SEPARATION_ORDER_GRANTED_COVER_LETTER_TEMPLATE_ID, SEPARATION_ORDER_GRANTED_SOLICITOR_COVERSHEET_DOCUMENT_NAME,
@@ -107,7 +115,8 @@ public class ConditionalOrderGrantedDocumentPack implements DocumentPack {
 
     private static final DocumentPackInfo APPLICANT_1_CONDITIONAL_ORDER_PACK = new DocumentPackInfo(
         ImmutableMap.of(
-            CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_1, Optional.of(CO_GRANTED_COVER_LETTER_TEMPLATE_ID)
+            CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_1, Optional.of(CO_GRANTED_COVER_LETTER_TEMPLATE_ID),
+            CONDITIONAL_ORDER_GRANTED, Optional.empty()
         ),
         ImmutableMap.of(
             CO_GRANTED_COVER_LETTER_TEMPLATE_ID, SEPARATION_ORDER_PRONOUNCED_DOCUMENT_NAME
@@ -116,7 +125,8 @@ public class ConditionalOrderGrantedDocumentPack implements DocumentPack {
 
     private static final DocumentPackInfo APPLICANT_2_CONDITIONAL_ORDER_PACK = new DocumentPackInfo(
         ImmutableMap.of(
-            CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_2, Optional.of(CO_GRANTED_COVER_LETTER_TEMPLATE_ID)
+            CONDITIONAL_ORDER_GRANTED_COVERSHEET_APP_2, Optional.of(CO_GRANTED_COVER_LETTER_TEMPLATE_ID),
+            CONDITIONAL_ORDER_GRANTED, Optional.empty()
         ),
         ImmutableMap.of(
             CO_GRANTED_COVER_LETTER_TEMPLATE_ID, SEPARATION_ORDER_PRONOUNCED_DOCUMENT_NAME
